+++ package.json
@@ -11,11 +11,12 @@
   },
   "lint-staged": {
     "*.{js,ts,tsx,jsx}": "eslint --fix"
- },
+  },
   "dependencies": {
     "next": "14.1.0",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "tailwind-merge": "^2.2.1"
   },
   "devDependencies": {
     "@types/node": "^20",
