+++ src/pages/index.tsx
@@ -1,5 +1,3 @@
-import Button from '@/components/common/Button';
-
 export default function Home() {
 	return <div>랜딩페이지</div>;
 }
